checking if the user exists in the deleteUser

+++ src/main/java/com/hotel_project/lake_side_hotel/service/UserServiceImpl.java
@@ -39,7 +39,10 @@ public class UserServiceImpl implements IUserService{
     @Transactional
     @Override
     public void deleteUser(String email) {
-        userRepository.deleteByEmail(email);
+        User theUser = getUser(email);
+        if (theUser != null) {
+            userRepository.deleteByEmail(email);
+        }
     }
 
     @Override
